requirements.txt
marshmallow==2.21.0
Flask-SQLAlchemy==2.4.4
Flask==1.1.2
Flask-rebar==1.12.2
python-dotenv==0.14.0
psycopg2-binary==2.8.6
uWSGI==2.0.19.1
dlib==19.21.0
face_recognition==1.3.0